feat: added utils dor DRY
